WIP: - refactoring the API now that we have LDIAG_SPEC: - some APIs only need a boolean debugflag (type: `l_ok` instead `l_int32`) so we should roll back those hard & fast initial edits. - Also code review and check the diagspec is used & passed along / propagated as intended everywhere:   + PIX produced from other PIX should always inherit their diagspec,   + while PIXA and other scenarios where multiple diagspecs may be incoming, the 'most prevalent' should be used,   i.e. the diagspec which is non-empty and has debug mode activated, i.e. debug mode in any diagspec should always take precedence. - also refactored the batch/item/step/... hierarchy code: we now provide a 10-levels deep arbitrary-use numeric hierarchy, so applications can use this mechanisms for other hierarchies. The filename/path formatting/generating code is still TODO, but the basics of this part are inspired by the NUMA APIs -- we chose NOT to use those as we're tracking/incrementing *integers* rather than floating point values, so we came up with our own StepsArray internal-use struct instead. - redesigned the FromAny APIs as it turned out that some of the 'valid' arguments COULD be NULL at run-time, so we don't really have a list of pointers terminated by a NULL sentinel. Hence we now use a argument count parameter for these APIs, so any usage can drop NULLs in when that makes the calling code simpler.
We got to `dewarpBuildPageModel` in the list; the rest is TODO...

+++ src/pix1.c
@@ -1988,7 +1988,7 @@ const PIXCMAP  *cmap;
             pixGetWidth(pix), pixGetHeight(pix), pixGetDepth(pix),
             pixGetSpp(pix));
     fprintf(fp, "    wpl = %d, data = %p, refcount = %d\n",
-            pixGetWpl(pix), pix->data, pix->refcount);
+            pixGetWpl(pix), pix->data, (int)pix->refcount);
     fprintf(fp, "    xres = %d, yres = %d\n", pixGetXRes(pix), pixGetYRes(pix));
     if ((cmap = pix->colormap) != NULL)
         pixcmapWriteStream(fp, cmap);
